@@ -1,10 +1,9 @@
 {
   "require": {
-    "php": "^7.4|^8.0",
+    "php": "^8.0",
     "ext-dom": "*",
     "ext-json": "*",
-    "nette/php-generator": "^3.6",
-    "symfony/polyfill-php80": "^1.27"
+    "nette/php-generator": "^4.1"
   },
   "autoload": {
     "psr-4": {
@@ -21,7 +20,10 @@
   "config": {
     "optimize-autoloader": true,
     "preferred-install": "dist",
-    "sort-packages": true
+    "sort-packages": true,
+    "platform": {
+      "php": "8.0"
+    }
   },
   "minimum-stability": "dev",
   "prefer-stable": true
