@@ -1,6 +1,4 @@
-<?php
-
-declare(strict_types=1);
+<?php declare(strict_types=1);
 
 /**
  * Handler.php
@@ -457,7 +455,7 @@ final class Handler
 			'attributes' => [
 				'maxlength'   => 255,
 				'value'       => $context['smm_skeleton']['min_php_version'],
-				'placeholder' => '7.4',
+				'placeholder' => '8.0',
 			],
 		];
 
@@ -789,7 +787,7 @@ final class Handler
 			$class = $namespace->addClass($classname);
 			$class->addComment('Generated by ' . SMM_NAME);
 
-			$context['smm_skeleton']['scheduled_tasks'][$id]['callable'] = "\\\\{$context['smm_skeleton']['author']}\\\\{$baseClassname}\\\\Tasks\\\\{$classname}::execute";
+			$context['smm_skeleton']['scheduled_tasks'][$id]['callable'] = "\\\\{$context['smm_skeleton']['author']}\\\\$baseClassname\\\\Tasks\\\\$classname::execute";
 
 			if (! empty($context['smm_skeleton']['use_final_class']))
 				$class->setFinal();
@@ -824,7 +822,7 @@ final class Handler
 			$class->setExtends('SMF_BackgroundTask');
 			$class->addComment('Generated by ' . SMM_NAME);
 
-			$context['smm_skeleton']['background_tasks'][$id]['callable'] = "\\\\{$context['smm_skeleton']['author']}\\\\{$baseClassname}\\\\Tasks\\\\{$classname}";
+			$context['smm_skeleton']['background_tasks'][$id]['callable'] = "\\\\{$context['smm_skeleton']['author']}\\\\$baseClassname\\\\Tasks\\\\$classname";
 
 			if (! empty($context['smm_skeleton']['use_final_class']))
 				$class->setFinal();
@@ -843,7 +841,7 @@ final class Handler
 			} else {
 				$regularity = $task['regularity'] == 1 ? 1 : 7;
 				$method->addBody("// Run task again if you need");
-				$method->addBody("\$regularity = {$regularity} * 24 * 60 * 60;" . PHP_EOL);
+				$method->addBody("\$regularity = $regularity * 24 * 60 * 60;" . PHP_EOL);
 				$method->addBody("return (bool) \$smcFunc['db_insert']('insert',");
 				$method->addBody("\t'{db_prefix}background_tasks',");
 				$method->addBody("\t[");
@@ -853,7 +851,7 @@ final class Handler
 				$method->addBody("\t\t'claimed_time' => 'int'");
 				$method->addBody("\t],");
 				$method->addBody("\t[");
-				$method->addBody("\t\t'\$sourcedir/{$baseClassname}/Tasks/{$filename}',");
+				$method->addBody("\t\t'\$sourcedir/$baseClassname/Tasks/$filename',");
 				$method->addBody("\t\t'\\\\' . self::class,");
 				$method->addBody("\t\t'',");
 				$method->addBody("\t\ttime() + \$regularity");
@@ -897,7 +895,7 @@ final class Handler
 			$method->addBody("\t\t'task_data'  => 'string'");
 			$method->addBody("\t],");
 			$method->addBody("\t[");
-			$method->addBody("\t\t'\$sourcedir/{$baseClassname}/Tasks/{$filename}',");
+			$method->addBody("\t\t'\$sourcedir/$baseClassname/Tasks/$filename',");
 			$method->addBody("\t\t'{$task['callable']}',");
 			$method->addBody("\t\t''");
 			$method->addBody("\t],");
@@ -912,7 +910,7 @@ final class Handler
 
 		$license = $context['smm_skeleton']['license_data'];
 
-		$file = new PhpFile;
+		$file = new PhpFile();
 
 		if (! empty($context['smm_skeleton']['use_strict_typing']))
 			$file->setStrictTypes();
@@ -928,7 +926,7 @@ final class Handler
 		$file->addComment('');
 		$file->addComment("@version " . $context['smm_skeleton']['version']);
 
-		return (new Printer)->printFile($file);
+		return (new Printer())->printFile($file);
 	}
 
 	private function fillConfigVars(Method $method, string $snake_name): void
@@ -995,16 +993,11 @@ final class Handler
 
 	private function getDefaultValue(array $option): string
 	{
-		switch ($option['type']) {
-			case 'int':
-				$default = (int) $option['default'];
-				break;
-			case 'float':
-				$default = (float) $option['default'];
-				break;
-			default:
-				$default = $option['default'];
-		}
+		$default = match ($option['type']) {
+			'int' => (int) $option['default'],
+			'float' => (float) $option['default'],
+			default => $option['default'],
+		};
 
 		return var_export($default, true);
 	}
